% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_addRT_mona.R
\name{addRT.mona}
\alias{addRT.mona}
\title{Add Retention Time to a mona.msp}
\usage{
addRT.mona(msp = "mspfile", prediction)
}
\arguments{
\item{msp}{A msp file downloaded from MoNA}

\item{prediction}{a dataframe from RT.spell function}
}
\value{
msp file with Predicted Retention time
}
\description{
Add Retention Time to a mona.msp
}
\examples{
\donttest{
addRT.mona(msp="MoNA-export-CASMI_2016.msp",mona_rt)}
}
