% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_cesc.R
\name{cesc}
\alias{cesc}
\title{Build a model to center and scale dataframe}
\usage{
cesc(db_rt)
}
\arguments{
\item{db_rt}{A dataframe previusly processed to clean na values}
}
\value{
model to use for center and scale a dataframe
}
\description{
Build a model to center and scale dataframe
}
\examples{
\donttest{
preProc <- cesc(db_rt)}
}
