% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_chemspace.R
\name{chem.space}
\alias{chem.space}
\title{Extract information from msp file downloaded from MoNA}
\usage{
chem.space(db_rt, target, title = "")
}
\arguments{
\item{db_rt}{A user library dataframe with chemical descriptors calculated}

\item{target}{Target database to be compared. Choice between: "ALL",  "HMDB",  "KNAPSACK", "CHEBI",
"DRUGBANK", "SMPDB", "YMDB", "T3DB", "FOODB", "NANPDB", "STOFF", "BMDB", "LIPIDMAPS", "URINE",
"SALIVA", "FECES", "ECMDB", "CSF", "SERUM", "PUBCHEM.1", "PLANTCYC", "UNPD", "BLEXP", "NPA", "COCONUT"}

\item{title}{A title for the plot. Example: Plasma - HMDB}
}
\value{
plot chemical space between user library and target database
}
\description{
Extract information from msp file downloaded from MoNA
}
\examples{
\donttest{
chem.space(db_rt,t="HMDB")}
}
