% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_xgboost.R
\name{fit.automl.h2o}
\alias{fit.automl.h2o}
\title{Build Automatic Machine Learning Model from H2O}
\usage{
\method{fit}{h2o.automl}(training)
}
\arguments{
\item{training}{A training dataset with calculated Chemical Descriptors}
}
\value{
Returns a trained model ready to predict
}
\description{
Build Automatic Machine Learning Model from H2O
}
\examples{
\donttest{
aml <- fit.automl.h2o(training)}
}
