% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_brnn.R
\name{fit.brnn}
\alias{fit.brnn}
\title{Build Bayesian Regularized Neural Network Model}
\usage{
\method{fit}{brnn}(t)
}
\arguments{
\item{t}{A training dataset with calculated Chemical Descriptors}
}
\value{
Returns a trained model ready to predict
}
\description{
Build Bayesian Regularized Neural Network Model
}
\examples{
\donttest{
brnn <- fit.brnn(training)}
}
