% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_keras.R
\name{fit.keras}
\alias{fit.keras}
\title{Build Keras model}
\usage{
\method{fit}{keras}(training, testing)
}
\arguments{
\item{training}{A training dataset with calculated Chemical Descriptors}

\item{testing}{A testing dataset with calculated Chemical Descriptors}
}
\value{
Returns a trained model ready to predict
}
\description{
Build Keras model
}
\examples{
\donttest{
keras <- fit.keras(training,testing)}
}
