% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_lighgbm.R
\name{fit.lightgbm}
\alias{fit.lightgbm}
\title{Build LightGBM model}
\usage{
\method{fit}{lightgbm}(training, testing)
}
\arguments{
\item{training}{A training dataset with calculated Chemical Descriptors}

\item{testing}{A testing dataset with calculated Chemical Descriptors}
}
\value{
Returns a trained model ready to predict
}
\description{
Build LightGBM model
}
\examples{
\donttest{
lightgbm <- fit.lightgbm(training,testing)}
}
