% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_getCD.R
\name{getCD}
\alias{getCD}
\title{Calculate Chemical Descriptors using RCDK}
\usage{
getCD(x)
}
\arguments{
\item{x}{A dataframe with 3 mandatory column: "Name", "InchIKey", "SMILES".}
}
\value{
A dataframe with calculated CD. Some compund can be missing if smiles is incorect or if molecule returns error in CDK
}
\description{
Calculate Chemical Descriptors using RCDK
}
\examples{
\donttest{
# RP and HILIC are previusly loaded dataset from excel with
# Name, InchIKey, SMILES and Retention Time
descs <- getCD(RP)
descs <- getCD(HILIC)}
}
