% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_getRT_smile.R
\name{getRT.smile}
\alias{getRT.smile}
\title{Calculate the Retention Time from a smiles entry}
\usage{
getRT.smile(
  smile = "O=C(O)C(NC(=O)C(N)C)CC(C)C",
  training,
  model = model,
  cesc = cesc
)
}
\arguments{
\item{smile}{A SMILE code}

\item{training}{A training dataset with calculated Chemical Descriptors used in the model}

\item{model}{A previusly computed model, like Xgboost, Keras etc}

\item{cesc}{A model for center and scale data calculated with cesc function}
}
\value{
Returns the Retention Time prediction in minutes
}
\description{
Calculate the Retention Time from a smiles entry
}
\examples{
\donttest{
smi <- "O=C2C(OC=3C=C(OC1OC(CO)C(O)C(O)C1(O))C=C(O)C2=3)=CC4=CC(O)=C(O)C(O)=C4"
getRT.smile(smile=smi,training,model=model_xgb)}
}
