% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HILIC}
\alias{HILIC}
\title{FiehnHilic positive dataset.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 970 rows and 4 columns.}
\source{
\href{http://mona.fiehnlab.ucdavis.edu/}{QTL Archive}
}
\usage{
data(HILIC)
}
\description{
FiehnHilic positive dataset.
}
\examples{
data(HILIC)
}
\references{
you have to chose one reference
}
\keyword{Fiehn}
\keyword{Hilic}
\keyword{Lab}
\keyword{datasets}
