% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_plotfeatures.R
\name{p.model.features}
\alias{p.model.features}
\title{Plot Feature Importance}
\usage{
p.model.features(model, mdl_type, h2o_number = 1)
}
\arguments{
\item{model}{A trained model (XGBoost, BRNN, RF, Lightgbm or autoML)}

\item{mdl_type}{The type of model trained (xgb, brnn, rf, lightgbm, h2o)}

\item{h2o_number}{The index of the autoML model to plot}
}
\value{
Returns one plot of the selected model
}
\description{
Plot Feature Importance
}
\examples{
\donttest{
p.model.features(model_xgb, xgb)}
}
