% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuction_plotmodel.R
\name{p.model}
\alias{p.model}
\title{Plot Model}
\usage{
p.model(t, m, title = "title", crh_leght = 16)
}
\arguments{
\item{t}{A dataset with calculated Chemical Descriptors like testing or training or whole dataset}

\item{m}{A trained model}

\item{title}{A title for the plot. Example: FiehnHilic Xgboost}

\item{crh_leght}{the chromatographic lengths in minutes}
}
\value{
Returns two plots of the selected model
}
\description{
Plot Model
}
\examples{
\donttest{
# db_rt is the dataset with calculated chemical descriptors
# and RT information coming from function getCD
set.seed(101)
inTraining <- createDataPartition(db_rt$XLogP, p = .8, list = FALSE)
training <- db_rt[ inTraining,]
testing  <- db_rt[-inTraining,]
model_xgb <- fit.xgboost(training)
p.model(testing,model_xgb)}
}
