% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_prep_mona.R
\name{prep.mona}
\alias{prep.mona}
\title{Extract information from msp file downloaded from MoNA}
\usage{
prep.mona(mspfile = "mona.msp")
}
\arguments{
\item{mspfile}{A msp file downloaded from MoNA}
}
\value{
dataframe with Name, InchIKey and SMILES
}
\description{
Extract information from msp file downloaded from MoNA
}
\examples{
\donttest{
mona <- prep.mona(msp=mona.msp)}
}
