% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_preprarewizard.R
\name{prep.wizard}
\alias{prep.wizard}
\title{Start parallel computing}
\usage{
prep.wizard()
}
\value{
numbers of cores
}
\description{
Start parallel computing
}
\examples{
\donttest{
prep_wizard()}
}
