% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_proc_data.R
\name{proc.data}
\alias{proc.data}
\title{Clean dataset removing NA value and Near zero variance columns.}
\usage{
proc.data(x)
}
\arguments{
\item{x}{A dataset with calculated Chemical Descriptors.}
}
\value{
Returns cleaned up dataset removing column with NA value and Near
zero variance.
}
\description{
Clean dataset removing NA value and Near zero variance columns.
}
\examples{
\donttest{
# descs is a previusly dataset with calculate chemical descriptors from
function getCD
db_rt <- proc.data(descs)}
}
