% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_retiplib.R
\docType{data}
\name{retip_lib_head}
\alias{retip_lib_head}
\title{All MS Fider library head.}
\format{An object of class \code{data.frame} with 321342 rows and 6 columns.}
\source{
\href{http://prime.psc.riken.jp/Metabolomics_Software/MS-FINDER/}{QTL Archive}
}
\usage{
data(retip_lib_head)
}
\description{
All MS Fider library head.
}
\examples{
data(retip_lib_head)
}
\references{
xxx
}
\keyword{MSFINDER}
\keyword{datasets}
\keyword{from}
