% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_RT_export.R
\name{RT.export}
\alias{RT.export}
\title{Export RT to external Mass Spectrometry Software}
\usage{
RT.export(data, program = "MSFINDER", pol = "pos")
}
\arguments{
\item{data}{The dataframe with RT prediction calculated you want to export}

\item{program}{The name of the software you want to use. Options are: MSDIAL, MSFINDER, AGILENT, THERMO, WATERS}

\item{pol}{Polarity you are going to use, could be positive "pos" or negative "neg"}
}
\value{
Write a file (txt or csv) formatted with external software requirements with Retention time calculated in the main folder
}
\description{
Export RT to external Mass Spectrometry Software
}
\examples{
\donttest{
export_rtp <- RT.export(data,program="MSDIAL",pol="pos")}
}
