% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_chemspace.R
\name{chem.space}
\alias{chem.space}
\title{Extract information from msp file downloaded from MoNA}
\usage{
chem.space(db_rt, t = "HMDB")
}
\arguments{
\item{db_rt}{A user library dataframe with chemical descriptors calculated}

\item{t}{Target database to be compared. Choice with: "ALL",  "HMDB",  "KNAPSACK",}
}
\value{
plot chemical space between user library and target database
}
\description{
Extract information from msp file downloaded from MoNA
}
\examples{
\donttest{
chem.space(db_rt,t="HMDB")}
}
