% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_automl.R
\name{fit.automl.h2o}
\alias{fit.automl.h2o}
\title{Build autoML Model}
\usage{
\method{fit}{automl.h2o}(training, max_models = 30)
}
\arguments{
\item{training}{A training dataset with calculated Chemical Descriptors}
}
\value{
Returns a trained model ready to predict
}
\description{
Build autoML Model
}
\examples{
\donttest{
aml <- fit.automl.h2o(training)}
}
