% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_chemspace.R
\name{p.model.features}
\alias{p.model.features}
\title{Plot Feature Importance}
\usage{
p.model.features(model = model_xgb, mdl_type = xgb)
}
\arguments{
\item{model}{A trained model}

\item{mdl_type}{The type of model trained (xgb, brnn, rf, lightgbm, h2o)}

\item{h2o_number}{The index of the autoML model to plot}
}
\value{
Returns a plot of feature importance (top20) of the model
}
\description{
Plot Feature Importance
}
\examples{
\donttest{
p.model.features(model = model_xgb, mdl_type = xgb)}
}